requirements.txt
langchain==0.2.16
langchain-community==0.2.16
pinecone-client==5.0.1
google-generativeai==0.7.2
sentence-transformers==3.0.1
gradio==4.44.0
python-dotenv==1.0.1
pyyaml==6.0.2